Add option to make control keys not trigger failure
Adds an additional configuration parameter, failonctrlkey. When set
to 0, slock will no longer change to the failure color if a control key
is pressed while the buffer is empty. This is useful if, for example,
you wake your monitor up by pressing a control key and don't want to
spoil the detection of failed unlocking attempts.